scripts/requirements.txt
boto3>=1.26.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
